added goal home page functionality
Added goal home page functionality.
Still trying to figure out how to add Sal's UI to the goal items
