+++ package.json
@@ -3,13 +3,16 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@devexpress/dx-react-core": "^2.7.2",
+    "@devexpress/dx-react-scheduler": "^2.7.2",
+    "@devexpress/dx-react-scheduler-material-ui": "^2.7.2",
     "@material-ui/core": "^4.11.0",
     "@material-ui/icons": "^4.9.1",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
     "axios": "^0.20.0",
-    "firebase": "^7.22.0",
+    "firebase": "^7.24.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-material-ui-calendar": "^0.2.3",
